refactor: use query params instead of route segment for editor
