+++ apps/dashboard/package.json
@@ -15,6 +15,7 @@
     "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
+    "@libsql/client": "^0.10.0",
     "@ogstudio/auth": "workspace:*",
     "@ogstudio/db": "workspace:*",
     "@radix-ui/themes": "^3.0.3",
