--- apps/dashboard/src/app/editor/[image]/page.tsx
@@ -1,11 +1,0 @@
-import { Editor } from "../../../components/Editor";
-
-export const metadata = {
-  title: "Editor - OG Studio",
-};
-
-export const dynamic = "force-static";
-
-export default function EditorPage() {
-  return <Editor />;
-}
