+++ apps/dashboard/src/components/Splash/MyImagesSplash.tsx
@@ -35,7 +35,7 @@ export function MyImagesSplash() {
               name: "New Image",
               id: createElementId(),
             };
-            router.push(`/editor/${image.id}`);
+            router.push(`/editor?i=${image.id}`);
             createImage(image);
           }}
         >
@@ -58,7 +58,7 @@ export function MyImagesSplash() {
             elements={
               JSON.parse(localStorage.getItem(image.id) ?? "[]") as OGElement[]
             }
-            href={`/editor/${image.id}`}
+            href={`/editor?i=${image.id}`}
             key={image.id}
             mockDynamicTexts
             name={image.name}
